requirements.txt
gql[all]
python-dotenv